remove githubPushStatus. use strings instead. use longs instead of ints

+++ app/repositories/models/GithubSubscriber.scala
@@ -4,4 +4,4 @@ case class GithubSubscriber(username: String,
                             repository: String,
                             token: String,
                             webhookUrl: String = null,
-                            id: Int = 0)
+                            id: Long = 0)
